develop

+++ src/main/kotlin/ru/foodbox/delivery/controllers/order/body/CreateOrderRequestBody.kt
@@ -3,6 +3,7 @@ package ru.foodbox.delivery.controllers.order.body
 import ru.foodbox.delivery.data.DeliveryType
 import ru.foodbox.delivery.services.dto.AddressDto
 import ru.foodbox.delivery.services.dto.OrderItemDto
+import java.math.BigDecimal
 
 class CreateOrderRequestBody(
     val deliveryType: DeliveryType,
@@ -10,6 +11,6 @@ class CreateOrderRequestBody(
     val comment: String?,
     val products: List<OrderItemDto>,
     val departmentId: Long,
-    val amount: Double,
-    val deliveryPrice: Double,
+    val amount: BigDecimal,
+    val deliveryPrice: BigDecimal,
 )
