+++ src/main/kotlin/ru/foodbox/delivery/services/model/DeliveryInfo.kt
@@ -1,6 +1,6 @@
 package ru.foodbox.delivery.services.model
 
 data class DeliveryInfo(
-    val deliveryPrice: Double,
-    val freeDeliveryPrice: Double?,
+    val deliveryPrice: Long,
+    val freeDeliveryPrice: Long?,
 )
